remove homepage and start with leaderboard
