+++ src/app/page.tsx
@@ -1,6 +1,17 @@
-import React from 'react';
-import HomePage from '@/components/home';
+import React from "react";
+import Game from "./game";
+import { getConfig } from "@/config/filesystem";
 
-export default function Home() {
-  return <HomePage />;
-}
+const page = async () => {
+  const config = await getConfig();
+
+  if (config instanceof Error) {
+    throw new Error(`Error loading config: ${config.message}`);
+  }
+
+  return (
+    <Game gameConfig={config} />
+  );
+};
+
+export default page;
